Studies from the Many Labs 2 project.

A subset of the data collected in the Many Labs 2 project which conducted replications of 28 classic and contemporary findings in psychology. The study examined the extent to which variability in replication success can be attributed to the study sample.

Usage

data(manylabs2018)

Format

A dataset with 1,414 rows and 23 variables.

lab: The lab which conducted the replication

es_id: Unique id for each effect size

yi_r: A numeric indicating the observed effect size, expressed in r

vi_r: A numeric indicating the variance on the observed effect size, expressed in r

yi_d: A numeric indicating the observed effect size, expressed in Cohen's d

vi_d: A numeric indicating the variance on the observed effect size, expressed in Cohen's d

ni: A numeric indicating the total sample size for the observed effect size

country: Country where the sample was collected

weird: Dummy variable encoding whether a country was classified as WEIRD; 0 = non-WEIRD, 1 = WEIRD

western: Dummy variable encoding a team judgment whether country was considered "western"

educated: Education score as measured by the Education Index

industrialized: Industrialization score as measured in the 2016 Industrial Development Report

rich: Dummy variable encoding whether a country is developed according to the 2014 World Economic Situation and Prospects Report; 0 = emerging or in transition, 1 = developed

democratic: The quality democracy in the corresponding country according to the 2015 Democracy Ranking Report. Higher scores indicate higher quality.

mean_weird_score: The arithmetic mean of the weird, western, educated, industrialized, and rich variables

online: Whether the study was replicated in a lab or online

analysis: Unique id for replicated study

Source

https://osf.io/ux3eh/

References

Klein, R. A., et al. (2018). Many Labs 2: Investigating variation in replicability across samples and settings. Advances in Methods and Practices in Psychological Science, 1(4), 443-490. (APS)