Studies on the association between goal orientation and performance adaptation

Results from 28 studies, including 86 effect sizes (r), on performance goal orientation and performance adaptation (Stasielowicz 2019) .

Usage

data("stasielowicz2019b")

Format

A data frame with 86 rows and 25 variables:

study_id: unique id for study

es_id: unique id for effect size

author: author of study

pub_year: publication year

pub_type: publication type (0 = journal article, 2 = book chapter, 3 = dissertation, 4 = master's thesis, 5 = bachelor's thesis, 6 = conference proceedings, 7 = report, 8 = other)

peer_review: whethere publication was peer-reviewed (0 = no, 1 = yes)

n: sample size of effect size

yi: effect size (r)

vi: sampling variance of effect size (SE^2)

pgo_type: the performance goal orientation that was assessed ("avoid", "prove", or "global")

adapt_measures: assessment method(s) of adaptation used in the study (1 = self-report, 2 = other people, 3 = objective, 4 = mixed)

adapt_method: assessment method of adaption (0 = subjective ratings, 1 = objective scores)

adapt_method_specific: specific assessment method of adaptation used for the particular effect size (1 = self-report, 2 = other people, 3 = objective)

go_measure: instrument used to assess goal orientation

financ_support: financial support (0 = no, 1 = yes)

age: mean age of sample

age_imputed: mean age of sample (imputed)

sex: sex of sample (1 = female sample, 2 = male sample, 3 = mixed sample)

perc_men: proportion of men in the sample

country: country where sampled was collected

sample: sample type (1 = students, 2 = employees, 3 = manager, 4 = mixed, 5 = other)

level: level (1 = individuals, 2 = team)

complexity_component: component complexity of the task (0 = relatively low, 1 = relatively high)

complexity_coordinative: coordinative complexity of the task (0 = relatively low, 1 = relatively high)

complexity_dynamic: which complexity aspect changed while completing task (0 = neither component nor coordinative, 1 = only component, 2 = only coordinative, 3 = both component and coordinative)

Details

Further details can be found at https://osf.io/szfwx/

References

Stasielowicz L (2019). “Goal orientation and performance adaptation: A meta-analysis.” Journal of Research in Personality, 82, 103847.