Studies on memory suppression and its deficiency in psychological disorders

Results from 25 studies, including 96 effect sizes (Cohen's d), on suppression-induced forgetting (Stramaccia et al. 2020) .

Usage

data("stramaccia2021")

Format

A data frame with 96 rows and 15 variables:

study_id: unique id for study

group_id: unique id for group

es_id: unique id for effect size

yi: effect size (Cohen's d)

vi: sampling variance for effect size (SE^2)

pub_year: year of publication

instructions: type of instructions given to participants to prevent retrieval ("aided", "direct", or "unspecified")

stimuli: type of stimuli ("pictures" or "words")

valence: valence of stimulus material (for the suppress targets only). One of "mixed", "negative", "neutral" or "positive"

tnttime: duration for which cues remained on the screen during the think/no-think phase (see paper for details)

repetitions: the number of times that participants encountered each cue in the think/no-think phase (see paper for details)

n: sample size

dv:

cluster: clusters based on clinical and sub-clinical conditions ("anxiety", "control", "depression", "mixed", or "repression")

group: clinical population ("CP") or healthy control ("HC")

Details

Further details can be found at https://osf.io/f89ur/

References

Stramaccia DF, Meyer A, Rischer KM, Fawcett JM, Benoit RG (2020). “Memory suppression and its deficiency in psychological disorders: A focused meta-analysis.” Journal of Experimental Psychology: General.